// frontend/src/client/models/Education.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Education = {
    institution?: (string | null);
    logo?: (string | null);
    degree?: (string | null);
    field_of_study?: (string | null);
    start_date?: (string | null);
    end_date?: (string | null);
};